+++ packages/office-ui-fabric-react/src/components/Button/DefaultButton/DefaultButton.tsx
@@ -4,7 +4,7 @@ import { BaseComponent, customizable, nullRender } from '../../../Utilities';
 import { IButtonProps } from '../Button.types';
 import { getStyles } from './DefaultButton.styles';
 
-@customizable('DefaultButton', ['theme'])
+@customizable('DefaultButton', ['theme', 'styles'])
 export class DefaultButton extends BaseComponent<IButtonProps, {}> {
   /**
    * Tell BaseComponent to bypass resolution of componentRef.