+++ packages/office-ui-fabric-react/src/components/Button/PrimaryButton/PrimaryButton.tsx
@@ -3,7 +3,7 @@ import { BaseComponent, customizable, nullRender } from '../../../Utilities';
 import { DefaultButton } from '../DefaultButton/DefaultButton';
 import { IButtonProps } from '../Button.types';
 
-@customizable('PrimaryButton', ['theme'])
+@customizable('PrimaryButton', ['theme', 'styles'])
 export class PrimaryButton extends BaseComponent<IButtonProps, {}> {
   /**
    * Set this BaseComponent._resolveComponentRef to false, bypassing resolution of componentRef.