+++ packages/office-ui-fabric-react/src/components/ComboBox/ComboBox.tsx
@@ -13,6 +13,7 @@ import {
   customizable,
   divProperties,
   findIndex,
+  focusAsync,
   getId,
   getNativeProps,
   shallowCompare
@@ -98,7 +99,7 @@ class ComboBoxOptionWrapper extends React.Component<IComboBoxOptionWrapperProps,
   }
 }
 
-@customizable('ComboBox', ['theme'])
+@customizable('ComboBox', ['theme', 'styles'])
 export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
   public static defaultProps: IComboBoxProps = {
     options: [],
@@ -134,18 +135,13 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
 
   // The current visible value sent to the auto fill on render
   private _currentVisibleValue: string | undefined;
-
   private _classNames: IComboBoxClassNames;
-
   private _isScrollIdle: boolean;
-
   private _hasPendingValue: boolean;
-
   private _scrollIdleTimeoutId: number | undefined;
-
   private _processingTouch: boolean;
-
   private _lastTouchTimeoutId: number | undefined;
+  private _processingExpandCollapseKeyOnly: boolean;
 
   // Determines if we should be setting
   // focus back to the input when the menu closes.
@@ -174,6 +170,7 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
 
     this._isScrollIdle = true;
     this._processingTouch = false;
+    this._processingExpandCollapseKeyOnly = false;
 
     const initialSelectedIndices: number[] = this._getSelectedIndices(props.options, selectedKeys);
 
@@ -243,7 +240,7 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
         this._autofill.current &&
         document.activeElement !== this._autofill.current.inputElement)
     ) {
-      this.focus();
+      this.focus(undefined /*shouldOpenOnFocus*/, true /*useFocusAsync*/);
     }
 
     // If we should focusAfterClose AND
@@ -424,11 +421,16 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
   }
 
   /**
-   * Set focus on the input
+   * @inheritdoc
    */
-  public focus = (shouldOpenOnFocus?: boolean): void => {
+  public focus = (shouldOpenOnFocus?: boolean, useFocusAsync?: boolean): void => {
     if (this._autofill.current) {
-      this._autofill.current.focus();
+      if (useFocusAsync) {
+        focusAsync(this._autofill.current);
+      } else {
+        this._autofill.current.focus();
+      }
+
       if (shouldOpenOnFocus) {
         this.setState({
           isOpen: true
@@ -1022,6 +1024,7 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
       onRenderList,
       calloutProps,
       dropdownWidth,
+      dropdownMaxWidth,
       onRenderLowerContent = this._onRenderLowerContent,
       useComboBoxAsMenuWidth
     } = props;
@@ -1044,6 +1047,7 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
             ? this._comboBoxWrapper.current.clientWidth + 2
             : dropdownWidth
         }
+        calloutMaxWidth={dropdownMaxWidth}
       >
         <div className={this._classNames.optionsContainerWrapper} ref={this._comboBoxMenu}>
           {(onRenderList as any)({ ...props }, this._onRenderList)}
@@ -1112,9 +1116,11 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
     const id = this._id;
     const isSelected: boolean = this._isOptionSelected(item.index);
     const optionStyles = this._getCurrentOptionStyles(item);
+    const optionClassNames = getComboBoxOptionClassNames(this._getCurrentOptionStyles(item));
     const checkboxStyles = () => {
       return optionStyles;
     };
+    const title = this._getPreviewText(item);
 
     const getOptionComponent = () => {
       return !this.props.multiSelect ? (
@@ -1133,10 +1139,11 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
           aria-selected={isSelected ? 'true' : 'false'}
           ariaLabel={this._getPreviewText(item)}
           disabled={item.disabled}
+          title={title}
         >
           {' '}
           {
-            <span ref={isSelected ? this._selectedElement : undefined}>
+            <span className={optionClassNames.optionTextWrapper} ref={isSelected ? this._selectedElement : undefined}>
               {onRenderOption(item, this._onRenderOptionContent)}
             </span>
           }
@@ -1155,6 +1162,7 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
           role="option"
           aria-selected={isSelected ? 'true' : 'false'}
           checked={isSelected}
+          title={title}
         >
           {onRenderOption(item, this._onRenderOptionContent)}
         </Checkbox>
@@ -1521,6 +1529,10 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
     const { disabled, allowFreeform, autoComplete } = this.props;
     const { isOpen, currentOptions, currentPendingValueValidIndexOnHover } = this.state;
 
+    // Take note if we are processing a altKey or metaKey keydown
+    // so that the menu does not collapse if no other keys are pressed
+    this._processingExpandCollapseKeyOnly = this._isExpandCollapseKey(ev);
+
     if (disabled) {
       this._handleInputWhenDisabled(ev);
       return;
@@ -1594,8 +1606,15 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
           index = this.state.currentOptions.length;
         }
 
-        if ((ev.altKey || ev.metaKey) && isOpen) {
-          this._setOpenStateAndFocusOnClose(!isOpen, true /* focusInputAfterClose */);
+        if (ev.altKey || ev.metaKey) {
+          // Close the menu if it is open and break so
+          // that the event get stopPropagation and prevent default.
+          // Otherwise, we need to let the event continue to propagate
+          if (isOpen) {
+            this._setOpenStateAndFocusOnClose(!isOpen, true /* focusInputAfterClose */);
+            break;
+          }
+
           return;
         }
 
@@ -1654,9 +1673,9 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
         }
 
         // If we get here and we got either and ALT key
-        // or meta key and we are current open, let's close the menu
-        if ((ev.altKey || ev.metaKey) && isOpen) {
-          this._setOpenStateAndFocusOnClose(!isOpen, true /* focusInputAfterClose */);
+        // or meta key, let the event propagate
+        if (ev.keyCode === KeyCodes.alt || ev.key === 'Meta' /* && isOpen */) {
+          return;
         }
 
         // If we are not allowing freeform and
@@ -1667,7 +1686,7 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
           break;
         }
 
-        // allow the key to propigate by default
+        // allow the key to propagate by default
         return;
     }
 
@@ -1675,12 +1694,22 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
     ev.preventDefault();
   };
 
+  private _isExpandCollapseKey(ev: React.KeyboardEvent<HTMLElement | Autofill>) {
+    return ev.which === KeyCodes.alt || ev.key === 'Meta';
+  }
+
   /**
    * Handle keyup on the input
    * @param ev - the keyboard event that was fired
    */
   private _onInputKeyUp = (ev: React.KeyboardEvent<HTMLElement | Autofill>): void => {
     const { disabled, allowFreeform, autoComplete } = this.props;
+    const isOpen = this.state.isOpen;
+
+    // If we get here and have only gotten the expand/collapse key
+    // and are processing the keyup of that event we should collapse
+    const shouldHandleKey = this._processingExpandCollapseKeyOnly && this._isExpandCollapseKey(ev);
+    this._processingExpandCollapseKeyOnly = false;
 
     if (disabled) {
       this._handleInputWhenDisabled(ev);
@@ -1693,13 +1722,14 @@ export class ComboBox extends BaseComponent<IComboBoxProps, IComboBoxState> {
         // make space expand/collapse the comboBox
         // and allow the event to propagate
         if (!allowFreeform && autoComplete === 'off') {
-          const isOpen = this.state.isOpen;
           this._setOpenStateAndFocusOnClose(!isOpen, !!isOpen);
           return;
         }
         break;
-
       default:
+        if (shouldHandleKey && isOpen) {
+          this._setOpenStateAndFocusOnClose(!isOpen, true /* focusInputAfterClose */);
+        }
         return;
     }