+++ packages/office-ui-fabric-react/src/components/Dialog/DialogContent.base.tsx
@@ -7,7 +7,7 @@ import { withResponsiveMode } from '../../utilities/decorators/withResponsiveMod
 
 const getClassNames = classNamesFunction<IDialogContentStyleProps, IDialogContentStyles>();
 
-@customizable('DialogContent', ['theme'])
+@customizable('DialogContent', ['theme', 'styles'])
 @withResponsiveMode
 export class DialogContentBase extends BaseComponent<IDialogContentProps, {}> {
   public static defaultProps: IDialogContentProps = {