+++ packages/office-ui-fabric-react/src/components/Shimmer/Shimmer.tsx
@@ -1,0 +1,6 @@
+import { styled } from '../../Utilities';
+import { IShimmerProps, IShimmerStyleProps, IShimmerStyles } from './Shimmer.types';
+import { getStyles } from './Shimmer.styles';
+import { ShimmerBase } from './Shimmer.base';
+
+export const Shimmer = styled<IShimmerProps, IShimmerStyleProps, IShimmerStyles>(ShimmerBase, getStyles);