+++ packages/office-ui-fabric-react/src/components/ThemeGenerator/ThemeRulesStandard.ts
@@ -180,25 +180,6 @@ export function themeRulesStandardCreator(): IThemeRules {
   slotRules[FabricSlots[FabricSlots.themeDark]].isCustomized = true;
   slotRules[FabricSlots[FabricSlots.themeDarker]].isCustomized = true;
 
-  // todo: can remove this once we remove these outdated slots from the product
-  const primaryBackground = 'primaryBackground';
-  slotRules[primaryBackground] = {
-    name: primaryBackground,
-    inherits: slotRules[FabricSlots[FabricSlots.white]],
-    isCustomized: false,
-    dependentRules: []
-  };
-  slotRules[FabricSlots[FabricSlots.white]].dependentRules.push(slotRules[primaryBackground]);
-
-  const primaryText = 'primaryText';
-  slotRules[primaryText] = {
-    name: primaryText,
-    inherits: slotRules[FabricSlots[FabricSlots.neutralPrimary]],
-    isCustomized: false,
-    dependentRules: []
-  };
-  slotRules[FabricSlots[FabricSlots.neutralPrimary]].dependentRules.push(slotRules[primaryText]);
-
   /*** SEMANTIC SLOTS */
   // create the SlotRule for a semantic slot
   function _makeSemanticSlotRule(semanticSlot: SemanticColorSlots, inheritedFabricSlot: FabricSlots): void {
@@ -216,8 +197,6 @@ export function themeRulesStandardCreator(): IThemeRules {
   // Basic simple slots
   _makeSemanticSlotRule(SemanticColorSlots.bodyBackground, FabricSlots.white);
   _makeSemanticSlotRule(SemanticColorSlots.bodyText, FabricSlots.neutralPrimary);
-  _makeSemanticSlotRule(SemanticColorSlots.disabledBackground, FabricSlots.neutralLighter);
-  _makeSemanticSlotRule(SemanticColorSlots.disabledText, FabricSlots.neutralTertiaryAlt);
 
   return slotRules;
 }