+++ packages/jest-serializer-merge-styles/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @uifabric/jest-serializer-merge-styles
 
-This log was last generated on Tue, 05 Jun 2018 00:44:30 GMT and should not be manually modified.
+This log was last generated on Thu, 14 Jun 2018 20:52:57 GMT and should not be manually modified.
+
+## 6.0.3
+Thu, 14 Jun 2018 20:52:57 GMT
+
+### Patches
+
+- Fix broken documentation links
 
 ## 6.0.0
 Tue, 05 Jun 2018 00:44:30 GMT